% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.buck.cv.R
\name{calc.buck.cv}
\alias{calc.buck.cv}
\title{Calculate Buckland CV}
\usage{
calc.buck.cv(cv)
}
\arguments{
\item{cv}{Coefficient of variation from survey}
}
\value{
A scalar used to calculate a 95\% confidence interval for survey data.
}
\description{
Calculates 95\% confidence intervals from survey CV
}
\references{
\insertRef{buckland1992}{pwsHerringBasa}
}
