% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ess.r
\name{calculate_ess}
\alias{calculate_ess}
\title{Calculate effective sample sizes for age composition data}
\usage{
calculate_ess(data, max_iter = 10, ...)
}
\arguments{
\item{data}{List of model data passed to \code{TMB::MakeADFun}}

\item{max_iter}{The maximum number of iterations in reweighting procedure}

\item{...}{Other parameters to be passed to \code{TMB::MakeADFun} and \code{stats::nlminb}}
}
\value{
List of effective sample sizes for age composition data and a convergence exit message.
}
\description{
Estimates effective sample sizes for age composition data in BASA using a
modified iterative reweighting procedure.
}
\details{
estimate age comps --> calculate ESS's --> estimate age comps --> repeat until convergence
}
\references{
\insertRef{muradian2017}{pwsHerringBasa}

\insertRef{stewart2014}{pwsHerringBasa}
}
