% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.biomass.traj.R
\name{compute.biomass.traj}
\alias{compute.biomass.traj}
\title{Compute Biomass Trajectory}
\usage{
compute.biomass.traj(dir_mcmc_out, nyr)
}
\arguments{
\item{dir_mcmc_out}{String giving directory containing mcmc outputs}

\item{nyr}{Total number of years in model time series}
}
\value{
A \link[tibble]{tibble} containing biomass (metric tons) over the full
model time series, as well as credible intervals and the probability that the
stock is below the harvest threshold (22000 tons or approximately 19958 metric tons).
}
\description{
Calculate model-estimated biomass credible intervals over full time-series
}
