% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.exploit.rate.R
\name{compute.exploit.rate}
\alias{compute.exploit.rate}
\title{Compute Exploitation Rate}
\usage{
compute.exploit.rate(dir_mcmc_out, nyr)
}
\arguments{
\item{dir_mcmc_out}{String giving directory containing mcmc outputs}

\item{nyr}{Total number of years in model time series}
}
\value{
A list of length 2 containing \link[tibble]{tibble}'s. The first
shows model-estimated exploitation rate with credible intervals over the full
model time series. The second gives the years in which the fishery was closed
(and thus, a known exploitation rate of 0).
}
\description{
Calculate model-estimated instantaneous exploitation rate with credible
intervals over full time-series
}
