% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.pfrb.posterior.R
\name{compute.pfrb.posterior}
\alias{compute.pfrb.posterior}
\title{Compute Pre-fishery Biomass Posterior}
\usage{
compute.pfrb.posterior(dir_mcmc_out, nyr)
}
\arguments{
\item{dir_mcmc_out}{String giving directory containing mcmc outputs}

\item{nyr}{Total number of years in model time series}
}
\value{
A list of length 3. The first (\code{biomass.df}) is a single-column
\code{data.frame} giving posterior distribution. The second (\code{biomass.quants})
is a numeric vector giving the .025, .5, and .975 quantiles of posterior
distribution. The third (\code{prob.below.threshold}) is a scalar giving the probability
that pre-fishery biomass lies below 20,000 metric tons (an approximation of the
22,000 short-ton harvest threshold).
}
\description{
Estimate pre-fishery biomass (metric tons) posterior distribution from MCMC runs
}
