% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.recruitment.R
\name{compute.recruitment}
\alias{compute.recruitment}
\title{Compute Recruitment Time Series}
\usage{
compute.recruitment(dir_mcmc_out, nyr, years)
}
\arguments{
\item{dir_mcmc_out}{String giving directory containing mcmc outputs}

\item{nyr}{Total number of years in model time series}

\item{years}{Integer vector of years in model}
}
\value{
A \link[tibble]{tibble} containing model-estimated age-3 recruits
(millions of fish) with credible intervals for full time series in model.
}
\description{
Calculate model-estimated age-3 recruits (in millions of fish) with credible
intervals.
}
