% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.retrospective.matrix.r
\name{compute.retrospective.matrix}
\alias{compute.retrospective.matrix}
\title{Compute retrospective matrix}
\usage{
compute.retrospective.matrix(n.peels, dir_mcmc_out)
}
\arguments{
\item{n.peels}{Integer giving the maximum number of years peeled in
retrospective analysis}

\item{dir_mcmc_out}{String giving directory containing base model mcmc outputs}
}
\value{
A matrix of median pre-fishery biomass for each year (rows) and
retrospective peel (columns)
}
\description{
Summarizes retrospective BASA model runs into a matrix for plotting in the
\code{plot_retrospective.r} script.
}
