% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.reader.R
\name{data.reader}
\alias{data.reader}
\title{Reads data files}
\usage{
data.reader(filename)
}
\arguments{
\item{filename}{Character string giving file path to data file.}
}
\value{
\code{R} list of data
}
\description{
Reads in .DAT and .CTL files created for ADMB .TPL files
}
