% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.post.pred.R
\name{generate.post.pred}
\alias{generate.post.pred}
\title{Generate posterior predictive intervals}
\usage{
generate.post.pred(
  fname,
  var,
  ess,
  years,
  nburn = 1,
  dist = "lognorm",
  mask = NA
)
}
\arguments{
\item{fname}{Character string giving file path to MCMC results for predicted survey data}

\item{var}{Vector of MCMC-derived variances for survey data (lognorm or negbin)}

\item{ess}{Vector of effective sample sizes (multinom only)}

\item{years}{Years fit by BASA model}

\item{nburn}{Length of burn-in period at beginning of markov chains}

\item{dist}{Assumed distribution of survey data. Supported inputs are
\code{lognorm}, \code{negbin}, and \code{multinom}}

\item{mask}{Years when survey data was not collected, avoids generating predictions
for those years}
}
\value{
A tibble of 50\% and 95\% posterior predictive intervals for survey data.
}
\description{
Generates posterior predictive intervals for log-normally, negative-binomially,
or multinomially
distributed survey data. Intervals calculated from distributions generated by
\link[pwsHerringBasa]{post.pred.lognorm}, \link[pwsHerringBasa]{post.pred.negbin},
and \link[pwsHerringBasa]{post.pred.multinom}.
}
