% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.admb.params.R
\name{init.admb.params}
\alias{init.admb.params}
\title{Set initial parameter values}
\usage{
init.admb.params(reps)
}
\arguments{
\item{reps}{The number of markov chains used in the model.}
}
\value{
List of named parameter values to initialize each markov chain in BASA.
}
\description{
Provides a set of initial parameter values for BASA's MCMC chains. Initial
values are provided in the \verb{PWS_ASA(par).ctl} file and random error is added
for intializing each markov chain.
}
