% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tmb_params.r
\name{init_tmb_params}
\alias{init_tmb_params}
\title{Set initial TMB parameter values}
\usage{
init_tmb_params(chains, lower, upper)
}
\arguments{
\item{chains}{The number of markov chains used in the model.}

\item{lower}{Vector of lower bounds for parameter initialization}

\item{upper}{Vector of upper bounds for parameter initialization}
}
\value{
List of named parameter values to initialize each markov chain in BASA.
}
\description{
Provides a set of initial parameter values for BASA's MCMC chains. Upper
and lower bounds for parameter values are provided as input and random
error is added for initializing each markov chain.
}
