% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listN.R
\name{listN}
\alias{listN}
\title{Create named list}
\usage{
listN(...)
}
\arguments{
\item{...}{Objects to include in named list}
}
\value{
A named list
}
\description{
A wrapper function for \link[base]{list} which names list elements using their
object name
}
