% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exploit_rate.R
\name{plot_exploit_rate}
\alias{plot_exploit_rate}
\title{Plot Exploitation Rate}
\usage{
plot_exploit_rate(df, zeros, years)
}
\arguments{
\item{df}{A data frame (or \link[tibble]{tibble}) returned by
\link[pwsHerringBasa]{compute.exploit.rate} containing model-estimated
exploitation rate time series with credible intervals}

\item{zeros}{A data frame (or \link[tibble]{tibble}) returned by
\link[pwsHerringBasa]{compute.exploit.rate} containing years in which the
fishery was closed (and thus, a known exploitation rate of 0).}

\item{years}{Integer vector of years in time series}
}
\value{
A \link[ggplot2]{ggplot} object showing model-estimated instantaneous
exploitation rate over the full model time series, as well as credible intervals
}
\description{
Wrapper function for plotting the model-estimated exploitation rate with
credible intervals.
}
