% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pfrb_posterior.R
\name{plot_pfrb_posterior}
\alias{plot_pfrb_posterior}
\title{Plot Pre-fishery biomass posterior}
\usage{
plot_pfrb_posterior(df, quants, prob, curr.year, font.size = 1)
}
\arguments{
\item{df}{A data frame returned by
\link[pwsHerringBasa]{compute.pfrb.posterior} containing model-fit
pre-fishery biomass posterior distribution}

\item{quants}{A numeric vector returned by
\link[pwsHerringBasa]{compute.pfrb.posterior} containing .025, .5, and .975
quantiles.}

\item{prob}{A scalar returned by
\link[pwsHerringBasa]{compute.pfrb.posterior} giving the probability
that pre-fishery biomass lies below the 22,000 short-ton (19,958 metric ton)
harvest threshold.}

\item{curr.year}{Integer giving the current year in the model run}

\item{font.size}{Plotting parameter passed to \link[ggplot2]{ggplot}}
}
\value{
A \link[ggplot2]{ggplot} object showing histogram of model-fit
pre-fishery biomass (metric tons)
}
\description{
Wrapper function for plotting the model-fit pre-fishery biomass (metric tons)
posterior distribution
}
