% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survey_fits.R
\name{plot_survey_fits}
\alias{plot_survey_fits}
\title{Plot survey fits}
\usage{
plot_survey_fits(
  fits,
  survey.data,
  y.max,
  title,
  ylabel = "",
  scale = 1,
  cvs = TRUE
)
}
\arguments{
\item{fits}{Model fits of survey data}

\item{survey.data}{Observed values from surveys}

\item{y.max}{Upper limit of y-axis, passed to \link[ggplot2]{ylim}}

\item{title}{Plot title, passed to \link[ggplot2]{labs}}

\item{ylabel}{y-axis label (usually name of survey), passed to
\link[ggplot2]{labs}}

\item{scale}{Scaling factor for dependent variable (survey data)}

\item{cvs}{Logical indicating whether survey CV's are included in plot}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Plots BASA model fits to survey values with posterior predictive interval with
\link[ggplot2]{ggplot}
}
