% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.pred.lognorm.R
\name{post.pred.lognorm}
\alias{post.pred.lognorm}
\title{Posterior predictive distribution (lognormal)}
\usage{
post.pred.lognorm(model.pred, sd)
}
\arguments{
\item{model.pred}{MCMC results for model-predicted values of annual survey data}

\item{sd}{Model-estimated standard deviation of survey data for each markov chain}
}
\value{
Matrix of posterior predictive distributions (rows) for annual survey data (columns)
}
\description{
Calculates posterior predictive distributions for log-normally distributed survey data.
A reparameterization explained
\href{https://msalganik.wordpress.com/2017/01/21/making-sense-of-the-rlnorm-function-in-r/}{here}
is used.
}
