% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.pred.negbin.R
\name{post.pred.negbin}
\alias{post.pred.negbin}
\title{Posterior predictive distribution (negative binomial)}
\usage{
post.pred.negbin(model.pred, sd)
}
\arguments{
\item{model.pred}{MCMC results for model-predicted values of annual survey data}

\item{sd}{Model-estimated standard deviation of survey data for each markov chain}
}
\value{
Matrix of posterior predictive distributions (rows) for annual survey data (columns)
}
\description{
Calculates posterior predictive distributions for negative-binomially distributed survey data.
}
