% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.admb.files.R
\name{read.admb.files}
\alias{read.admb.files}
\title{Read data files input to ADMB}
\usage{
read.admb.files(model.dir = ".")
}
\arguments{
\item{model.dir}{String giving model directory}
}
\value{
Named list of model data.
}
\description{
A simple wrapper for reading in all of the model inputs supplied to ADMB.
}
