% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.biomass.estimates.R
\name{read.biomass.estimates}
\alias{read.biomass.estimates}
\title{Read in Biomass Estimates}
\usage{
read.biomass.estimates(dir_mcmc_out, nyr = NA)
}
\arguments{
\item{dir_mcmc_out}{String giving directory containing mcmc outputs}

\item{nyr}{Total number of years in model time series}
}
\value{
A data frame containing MCMC results (rows) for annual biomass
(columns, in metric tons)
}
\description{
Read in model-estimated biomass over full time-series
}
