% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.data.files.R
\name{read.data.files}
\alias{read.data.files}
\title{Read Data Files}
\usage{
read.data.files(dat.dir)
}
\arguments{
\item{dat.dir}{A character string giving the directory containg the \code{.dat},
\code{.ctl}, and \code{.txt} data files}
}
\value{
A named list with an element corresponding to each data file.
}
\description{
Read in the five data files containing survey data.
}
