% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.exploit.rates.R
\name{read.exploit.rates}
\alias{read.exploit.rates}
\title{Read in Exploitation Rates}
\usage{
read.exploit.rates(dir_mcmc_out, nyr = NA)
}
\arguments{
\item{dir_mcmc_out}{String giving directory containing mcmc outputs}

\item{nyr}{Total number of years in model time series}
}
\value{
A data frame containing MCMC results (rows) for annual estimated
instantaneous exploitation rates (columns)
}
\description{
Read in model-estimated exploitation rates over full time-series
}
