% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.basa.R
\name{run.basa}
\alias{run.basa}
\title{Run BASA model}
\usage{
run.basa(
  dir_model,
  n.samples = 2000,
  n.warmup = 700,
  n.time = 5,
  n.chains = 4,
  control = NULL
)
}
\arguments{
\item{dir_model}{Relative path to model}

\item{n.samples}{Number of samples to draw from MCMC iterations}

\item{n.warmup}{Number of MCMC burn-in iterations}

\item{n.time}{Number of minutes after which the model will quit running}

\item{n.chains}{Number of MCMC chains to run}

\item{control}{List of MCMC controls passed to \code{rstan}}
}
\value{
Metadata from model fit in ADMB. Model outputs are saved to \verb{model/mcmc_out/}
and \code{model/rep_out} subdirectories.
}
\description{
Compiles ADMB code, reads in data, calculates effective sample sizes, and
executes BASA model using NUTS sampling algorithm. This function \emph{should}
work for linux and mac, but has only been tested on windows.
}
