% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.basa.retro.r
\name{run.basa.retro}
\alias{run.basa.retro}
\title{Run BASA on peeled data}
\usage{
run.basa.retro(i, dir_model, dir_retro)
}
\arguments{
\item{i}{Number of years to peel}

\item{dir_model}{Character string giving model directory}

\item{dir_retro}{Character string giving retrospective directory}
}
\value{
A copy of (peeled) BASA outputs in \code{dir_retro}.
}
\description{
Helper function to run BASA on peeled data for retrospective analysis. Peeled
data is returned by \link[pwsHerringBasa]{write_retro_data}.
}
