% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_evenness.r
\name{sw_evenness}
\alias{sw_evenness}
\title{Shannon-Weiner Evenness}
\usage{
sw_evenness(naa)
}
\arguments{
\item{naa}{A model-estimated numbers-at-age vector for a single year}
}
\value{
A scalar, J, indicating the Shannon-Weiner evenness index score for
a single year's estimated age composition
}
\description{
Calculates the Shannon-Weiner evenness index score for age composition model
estimates.
}
\references{
\insertRef{shannonweaver1949}{pwsHerringBasa}
}
