% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.r
\name{write_report}
\alias{write_report}
\title{Write model report}
\usage{
write_report(
  obj,
  par,
  file,
  llik = NULL,
  derived = NULL,
  survey = NULL,
  forecast = NULL
)
}
\arguments{
\item{obj}{A list object returned by \code{MakeADFun}}

\item{par}{A list of parameter values to generate report}

\item{file}{File path to write report}

\item{llik}{Character vector of names of log-likelihood components}

\item{derived}{Character vector of names of derived quantities}

\item{survey}{Character vector of names of survey quantities}

\item{forecast}{Character vector of names of forecast quantities}
}
\description{
Utility function for writing model estimates of estimated parameters and
derived quantities.
}
