% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_retro_dat.r
\name{write_retro_dat}
\alias{write_retro_dat}
\title{Write peeled data for retrospective analysis}
\usage{
write_retro_dat(dat.files, years.to.remove, dir_retro)
}
\arguments{
\item{dat.files}{an R object containing BASA data files (returned by
\link[pwsHerringBasa]{read.data.files})}

\item{years.to.remove}{Integer giving the number of years to peel}

\item{dir_retro}{Character string giving retrospective directory}
}
\value{
A copy of (peeled) BASA inputs in \code{dir_retro}.
}
\description{
Peels terminal years off of BASA input data time series and writes to \code{.dat}
files. This peeled data is then fed to \link[pwsHerringBasa]{run.basa.retro}
which copies it to the correct directory and runs BASA.
}
