% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tmb_params.r
\name{init_tmb_params}
\alias{init_tmb_params}
\title{Set initial TMB parameter values}
\usage{
init_tmb_params(start, lower, upper, chains, seed)
}
\arguments{
\item{start}{Vector of starting values for parameters}

\item{lower}{Vector of lower bounds for parameter initialization}

\item{upper}{Vector of upper bounds for parameter initialization}

\item{chains}{Integer number of markov chains for which initial values are desired}

\item{seed}{Integer seed for generating random parameter starting values}
}
\value{
List of named parameter values to initialize each markov chain in BASA.
}
\description{
Provides a set of initial parameter values for BASA's MCMC chains. Upper
and lower bounds for parameter values are provided as input and random
error is added for initializing each markov chain.
}
